Go rid of uneccessary function. Made calculation a member function so that it no longer needs to pass parameters it just takes from whatever member is called from and returns their per-cycle exp
